# docs/swagger.yaml
definitions:
...
  api.AddressesRequest:
    items:
      properties:
        address:
          type: string
        coin:
          type: integer
      type: object
    type: array
...
  blockatlas.CollectionPage:
    items:
      properties:
        address:
          type: string
        category_address:
          type: string
        coin:
          type: integer
        description:
          type: string
        external_link:
          type: string
        id:
          type: string
        image_url:
          type: string
        name:
          type: string
        nft_version:
          type: string
        slug:
          type: string
        symbol:
          type: string
        total:
          type: integer
        type:
          type: string
      type: object
    type: array
...
  blockatlas.DelegationResponse:
    properties:
      address:
        type: string
      balance:
        type: string
      coin:
        $ref: '#/definitions/coin.ExternalCoin'
        type: object
      delegations:
        $ref: '#/definitions/blockatlas.DelegationsPage'
        type: object
      details:
        $ref: '#/definitions/blockatlas.StakingDetails'
        type: object
    type: object
  blockatlas.DelegationsBatchPage:
    items:
      properties:
        address:
          type: string
        balance:
          type: string
        coin:
          $ref: '#/definitions/coin.ExternalCoin'
          type: object
        delegations:
          $ref: '#/definitions/blockatlas.DelegationsPage'
          type: object
        details:
          $ref: '#/definitions/blockatlas.StakingDetails'
          type: object
      type: object
    type: array
  blockatlas.DelegationsPage:
    items:
      properties:
        delegator:
          $ref: '#/definitions/blockatlas.StakeValidator'
          type: object
        metadata:
          type: object
        status:
          type: string
        value:
          type: string
      type: object
    type: array
  blockatlas.DocsResponse:
    properties:
      docs:
        type: object
    type: object
...
  blockatlas.StakeValidator:
    properties:
      details:
        $ref: '#/definitions/blockatlas.StakingDetails'
        type: object
      id:
        type: string
      info:
        $ref: '#/definitions/blockatlas.StakeValidatorInfo'
        type: object
      status:
        type: boolean
    type: object
  blockatlas.StakeValidatorInfo:
    properties:
      description:
        type: string
      image:
        type: string
      name:
        type: string
      website:
        type: string
    type: object
  blockatlas.StakingDetails:
    properties:
      locktime:
        type: integer
      minimum_amount:
        type: string
      reward:
        $ref: '#/definitions/blockatlas.StakingReward'
        type: object
      type:
        type: string
    type: object
  blockatlas.StakingReward:
    properties:
      annual:
        type: number
    type: object
...
  blockatlas.TxOutput:
    properties:
      address:
        type: string
      value:
        type: string
    type: object
  blockatlas.TxPage:
    items:
      properties:
        block:
          description: Height of the block the transaction was included in
          type: integer
        coin:
          description: SLIP-44 coin index of the platform
          type: integer
        date:
          description: Unix timestamp of the block the transaction was included in
          type: integer
        direction:
          description: Transaction Direction
          type: string
        error:
          description: |-
            Empty if the transaction was successful,
            else error explaining why the transaction failed (optional)
          type: string
        fee:
          description: Transaction fee (native currency)
          type: string
        from:
          description: Address of the transaction sender
          type: string
        id:
          description: Unique identifier
          type: string
        inputs:
          description: Input addresses
          items:
            $ref: '#/definitions/blockatlas.TxOutput'
          type: array
        memo:
          description: Meta data object
          type: string
        metadata:
          type: object
        outputs:
          description: Output addresses
          items:
            $ref: '#/definitions/blockatlas.TxOutput'
          type: array
        sequence:
          description: Transaction nonce or sequence
          type: integer
        status:
          description: Status of the transaction
          type: string
        to:
          description: Address of the transaction recipient
          type: string
        type:
          description: Type of metadata
          type: string
      type: object
    type: array
  coin.ExternalCoin:
    properties:
      coin:
        type: integer
      decimals:
        type: integer
      name:
        type: string
      symbol:
        type: string
    type: object
  ginutils.ApiError:
    properties:
      status_code:
        type: integer
      status_message:
        type: string
    type: object
info:
  contact: {}
  license: {}
paths:
...
  /v2/{coin}/collections/{address}:
    get:
      consumes:
      - application/json
      description: Get all collections from the address
      operationId: collections
      parameters:
      - default: ethereum
        description: the coin name
        in: path
        name: coin
        required: true
        type: string
      - default: 0x5574Cd97432cEd0D7Caf58ac3c4fEDB2061C98fB
        description: the query address
        in: path
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/blockatlas.CollectionPage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ginutils.ApiError'
      summary: Get Collections
      tags:
      - platform
      - collection
  /v2/{coin}/collections/{owner}/collection/{collection_id}:
    get:
      consumes:
      - application/json
      description: Get a collection from the address
      operationId: collection
      parameters:
      - default: ethereum
        description: the coin name
        in: path
        name: coin
        required: true
        type: string
      - default: 0x0875BCab22dE3d02402bc38aEe4104e1239374a7
        description: the query address
        in: path
        name: owner
        required: true
        type: string
      - default: 0x06012c8cf97bead5deae237070f9587f8e7a266d
        description: the query collection
        in: path
        name: collection_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/blockatlas.CollectionPage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ginutils.ApiError'
      summary: Get Collection
      tags:
      - platform
      - collection
  /v2/{coin}/staking/delegations/{address}:
    get:
      consumes:
      - application/json
      description: Get stake delegations from the address
      operationId: delegations
      parameters:
      - default: tron
        description: the coin name
        in: path
        name: coin
        required: true
        type: string
      - default: TPJYCz8ppZNyvw7pTwmjajcx4Kk1MmEUhD
        description: the query address
        in: path
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/blockatlas.DelegationResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ginutils.ApiError'
      summary: Get Stake Delegations
      tags:
      - platform
      - staking
  /v2/{coin}/staking/validators:
    get:
      consumes:
      - application/json
      description: Get validators from the address
      operationId: validators
      parameters:
      - default: cosmos
        description: the coin name
        in: path
        name: coin
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/blockatlas.DocsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ginutils.ApiError'
      summary: Get Validators
      tags:
      - platform
      - staking
  /v2/{coin}/tokens/{address}:
    get:
      consumes:
      - application/json
      description: Get tokens from the address
      operationId: tokens
      parameters:
      - default: ethereum
        description: the coin name
        in: path
        name: coin
        required: true
        type: string
      - default: 0x5574Cd97432cEd0D7Caf58ac3c4fEDB2061C98fB
        description: the query address
        in: path
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/blockatlas.CollectionPage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ginutils.ApiError'
      summary: Get Tokens
      tags:
      - platform
      - token
  /v2/{coin}/transactions/{address}:
    get:
      consumes:
      - application/json
      description: Get transactions from the address
      operationId: tx_v2
      parameters:
      - default: tezos
        description: the coin name
        in: path
        name: coin
        required: true
        type: string
      - default: tz1WCd2jm4uSt4vntk4vSuUWoZQGhLcDuR9q
        description: the query address
        in: path
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/blockatlas.TxPage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ginutils.ApiError'
      summary: Get Transactions
      tags:
      - platform
      - tx
  /v2/collectibles/categories:
    post:
      consumes:
      - application/json
      description: Get collection categories
      operationId: collection_categories
      parameters:
      - default: '{"60": ["0xb3624367b1ab37daef42e1a3a2ced012359659b0"]}'
        description: Payload
        in: body
        name: data
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/blockatlas.DocsResponse'
      summary: Get list of collections from a specific coin and addresses
      tags:
      - Collectibles
...
  /v2/staking/delegations:
    post:
      consumes:
      - application/json
      description: Get Stake Delegations for multiple coins
      operationId: batch_delegations
      parameters:
      - description: Validators addresses and coins
        in: body
        name: delegations
        required: true
        schema:
          $ref: '#/definitions/api.AddressesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/blockatlas.DelegationsBatchPage'
      summary: Get Multiple Stake Delegations
      tags:
      - platform
      - staking
  /v2/staking/list:
    post:
      consumes:
      - application/json
      description: Get Stake Delegations for multiple coins
      operationId: batch_delegations
      parameters:
      - description: Validators addresses and coins
        in: body
        name: delegations
        required: true
        schema:
          $ref: '#/definitions/api.AddressesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/blockatlas.DelegationsBatchPage'
      summary: Get Multiple Stake Delegations
      tags:
      - platform
      - staking
swagger: "2.0"